fix: resolved warnings under -Wconversion on Clang14

+++ replacement/drrip/drrip.cc
@@ -3,7 +3,7 @@
 #include <cassert>
 #include <utility>
 
-drrip::drrip(CACHE* cache) : replacement(cache), NUM_SET(cache->NUM_SET), NUM_WAY(cache->NUM_WAY), rrpv(NUM_SET * NUM_WAY)
+drrip::drrip(CACHE* cache) : replacement(cache), NUM_SET(cache->NUM_SET), NUM_WAY(cache->NUM_WAY), rrpv(static_cast<std::size_t>(NUM_SET * NUM_WAY))
 {
   // randomly selected sampler sets
   std::size_t rand_seed = 1103515245 + 12345;
